% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ENAHO2012}
\alias{ENAHO2012}
\title{Data de la ENAHO 2012.}
\format{Dataframe con la data ENAHO 2012.}
\usage{
ENAHO2012
}
\description{
Para el detalle y diccionario de datos ir a la direccion:http://www.inec.go.cr/anda4/index.php/catalog/130/accesspolicy
}
\details{
Data de la Encuesta Nacional de Hogares 2012
}
\keyword{datasets}

