% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ENAHO2013}
\alias{ENAHO2013}
\title{ENAHO 2013 data.}
\format{Dataframe with ENAHO 2013 data.}
\usage{
ENAHO2013
}
\description{
For the details and data dictionaries go to:http://www.inec.go.cr/anda4/index.php/catalog/137
}
\details{
National Survey of homes,Costa Rica 2013
}
\keyword{datasets}

