% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ENAHO_2013}
\alias{ENAHO_2013}
\title{Cargar data de la ENAHO 2013}
\format{Dataframe con la data ENAHO 2013.}
\usage{
ENAHO_2013()
}
\arguments{
\item{No}{requiere argumentos.}
}
\description{
Para el detalle y diccionario de datos ir a la direccion:http://www.inec.go.cr/anda4/index.php/catalog/130/accesspolicy
}
\details{
Carga en un dataframe la data de la ENAHO 2013
}

