% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nsize}
\alias{nsize}
\title{Calculo de tamaños muestrales}
\usage{
nsize(x, r, alpha = 0.05)
}
\arguments{
\item{df:}{Data frame con datos para calcular media y variancia.}

\item{x:}{Columna base para calcular tamaño de muestra.}

\item{r:}{Error relativo.}

\item{alpha:}{Riesgo alpha de que el error real supere r.}
}
\value{
lista con n0 y n ajustado.
}
\description{
Basado en slides del curso Topicos de Muestreo.Universidad de Costa Rica.
}
\details{
Funcion para generar una muestra con control de error absoluto o relativo.
}

