% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sys.sample}
\alias{sys.sample}
\title{Systematic Sample selection}
\usage{
sys.sample(N, n)
}
\arguments{
\item{N:}{Total elements of population.}

\item{n:}{Sample size.}
}
\value{
Data object with the selected elements.
}
\description{
Select a sample with systematic algorithm.
}
\references{
Function taked from http://fisher.utstat.utoronto.ca/~mahinda/stac50/Rcode_sysSampling.txt
}

