% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ENAHO2012}
\alias{ENAHO2012}
\title{ENAHO 2012 data.}
\format{Dataframe with ENAHO 2013 data.}
\usage{
ENAHO2012
}
\description{
For the details and data dictionaries go to:http://www.inec.go.cr/anda4/index.php/catalog/130/accesspolicy
}
\details{
National Survey of homes,Costa Rica 2012
}
\keyword{datasets}

