% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ENAHO2013}
\alias{ENAHO2013}
\title{Data de la ENAHO 2013.}
\format{Dataframe con la data ENAHO 2013.}
\usage{
ENAHO2013
}
\description{
Para el detalle y diccionario de datos ir a la direccion:http://www.inec.go.cr/anda4/index.php/catalog/130/accesspolicy
}
\details{
Data de la Encuesta Nacional de Hogares 2013
}
\keyword{datasets}

