% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nsize}
\alias{nsize}
\title{Sample size calculation}
\usage{
nsize(x, r = 0, abs = 0, alpha = 0.05)
}
\arguments{
\item{x:}{Base vector with data for calculate sample size.}

\item{r:}{Relative error.}

\item{abs:}{Absolute error.}

\item{alpha:}{Alpha risk that the actual error exceeds r.}
}
\value{
list with n0 and adjust n.
}
\description{
Calculate sample sizes(absolute and relative error for continuous or dichotomous variables).
}
\references{
Based in the slides of Sampling Topics course.University of Costa Rica.
}

