% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nstrata}
\alias{nstrata}
\title{Calculo de tamaños muestrales en estratos}
\usage{
nstrata(n, wh, sh = NULL, ch = NULL, method = "proportional")
}
\arguments{
\item{n:}{Tamaño de la muestra total.}

\item{wh:}{Objeto on las proporciones por estrato.}

\item{sh:}{Desviación de cada estrato.}

\item{ch:}{Costo por entrevista .}

\item{method:}{Método de asignación de la muestra a los estratos.}
}
\value{
Objeto con muestra distribuida por estrato.
}
\description{
Funcion para distribuir la muestra en los estratos.
}
\references{
Basado en slides del curso Topicos de Muestreo.Universidad de Costa Rica.
}

