% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sys.sample}
\alias{sys.sample}
\title{Genera una muestra sistematica}
\usage{
sys.sample(N, n)
}
\arguments{
\item{N:}{Total de elementos de la Poblacion.}

\item{n:}{Elementos a seleccionar en la muestra.}
}
\value{
objeto con los elementos seleccionados.
}
\description{
Funcion para generar una muestra con algoritmo sistematico de n elementos de un total de N.
}
\references{
Funcion tomada de http://fisher.utstat.utoronto.ca/~mahinda/stac50/Rcode_sysSampling.txt
}

